Skip tests that won't work on Python 2.6 (fixes #20)
Some tests do not succeed due to the age of the Unicode data
contained in the Python 2.6 distribution. This allows tests to
pass by skipping them. A better long term solution is to pull
Unicode directionality data that matches the Unicode version
for which the other IDNA properties are derived, rather than
using the unicodedata module that comes with Python.